Extract project Python artifact lifecycle
