+++ cli/python/base_setup/tests/test_python_version_artifacts.py
@@ -6,8 +6,8 @@ from pathlib import Path
 from unittest import mock
 
 from base_setup import artifacts
-from base_setup.artifacts import ProjectRuntimeConfig
 from base_setup.errors import ArtifactError
+from base_setup.python_artifacts import ProjectRuntimeConfig
 from base_setup.python_policy import PythonInterpreter
 from base_setup.registry import get_artifact_definition
 from base_setup.tests.helpers import fake_context
@@ -23,11 +23,11 @@ class PythonVersionArtifactTests(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tmpdir:
             venv_dir = Path(tmpdir) / "demo" / ".venv"
             python_bin = Path(tmpdir) / "python3.12"
-            with mock.patch("base_setup.artifacts.project_venv_dir", return_value=venv_dir), mock.patch(
-                "base_setup.artifacts.python_artifact_installed",
+            with mock.patch("base_setup.python_artifacts.project_venv_dir", return_value=venv_dir), mock.patch(
+                "base_setup.python_artifacts.python_artifact_installed",
                 return_value=False,
             ), mock.patch(
-                "base_setup.artifacts.resolve_python_interpreter",
+                "base_setup.python_artifacts.resolve_python_interpreter",
                 return_value=PythonInterpreter(path=python_bin, version=(3, 12)),
             ), mock.patch(
                 "base_setup.process.run_command"
@@ -70,8 +70,8 @@ class PythonVersionArtifactTests(unittest.TestCase):
             python_bin.write_text("#!/bin/sh\nprintf '3.13\\n'\n", encoding="utf-8")
             python_bin.chmod(0o755)
 
-            with mock.patch("base_setup.artifacts.project_venv_dir", return_value=venv_dir), mock.patch(
-                "base_setup.artifacts.python_artifact_installed",
+            with mock.patch("base_setup.python_artifacts.project_venv_dir", return_value=venv_dir), mock.patch(
+                "base_setup.python_artifacts.python_artifact_installed",
                 return_value=False,
             ), mock.patch("base_setup.process.run_command") as run_command:
                 with self.assertRaisesRegex(ArtifactError, "uses Python 3.13"):
@@ -84,4 +84,3 @@ class PythonVersionArtifactTests(unittest.TestCase):
                     )
 
         run_command.assert_not_called()
-
